Updated files and added oledDisplay cpp startup file
Updated the header file with OLED libraries. There are two libraries that can be used interchangeably. One is text only and uses less program storage space and dynamic memory. The other is the Adafruit library that supports shapes and bitmaps. This must be defined inside components.h. The OLED is only setup there. So you still have to must call the normal library display.TheCommand.